netflixa_agent/requirements.txt
langgraph==0.2.0
langchain==0.2.0
langchain-groq==0.1.0
snowflake-connector-python==3.7.0
pandas==2.1.4
python-dotenv==1.0.0
streamlit==1.29.0
chromadb==0.4.22
sentence-transformers==2.2.2
numpy==1.24.3
plotly==5.17.0
